Fix D3D12 error
Error was: D3D12 reset command allocator failed with E_FAIL (0x80004005)

This error happened on 2 of the 3 machines I tested. I believe it was due to too many calls to GPUCanvasContext.configure().
Actually i found out that there was no need to systematically call this after a canvas resize, because when assigning to canvas.width/height, the last configuration was reused.